feat: add abort option to html requests

+++ Day 83 - Portfolio Website/main.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template, request
 from flask_sqlalchemy import SQLAlchemy
-from flask_restful import Api, Resource, fields, marshal
+from flask_restful import Api, Resource, fields, marshal, abort
 
 app = Flask(__name__)
 api = Api(app)
@@ -19,7 +19,7 @@ class Technology(db.Model):
 class SubTechnology(db.Model):
     __tablename__ = 'subtechnologies'
     id = db.Column(db.Integer, primary_key=True)
-    technology_id = db.Column(db.String, db.ForeignKey('technologies.id'))
+    # technology_id = db.Column(db.String, db.ForeignKey('technologies.id'))  # create a technology first if doesn't exist exception
     technology_name = db.Column(db.String, db.ForeignKey('technologies.technology_name'))
     subtechnology_name = db.Column(db.String(50), nullable=False)
 
@@ -57,6 +57,13 @@ def marshal_wo_null(content):
 
     return content_marshal_wo_null
 
+def abort_if_exist(content, checked_var):
+    try:
+        if checked_var in content[0].values(): # is there a possibility to do it differently? explore content[0].values()
+            abort(409, message=f'{checked_var} is already in database')
+    except IndexError:
+        pass
+
 
 class GetContent(Resource):
     def get(self):
@@ -64,27 +71,35 @@ class GetContent(Resource):
         content_type = args['content']
         content_id = args['id']
         got_content = globals().get(content_type).query.filter_by(id=content_id).first()  # !
-
-        return marshal_wo_null(got_content), 200
+        if not got_content:
+            abort(404, message='Couldn\'t find requested content')
+        else:
+            return marshal_wo_null(got_content), 200
 
 
 class PostContent(Resource):
     def post(self):
-        content = request.form
-        if content['content'] == 'technology':
+        args = request.form
+        content_type = args['content']
+        all_content = GetAllContent().get()
+
+        if content_type == 'technology':
+            table = all_content['Technologies']
+            abort_if_exist(table, args['technology_name'])
             new_content = Technology(
-                technology_name=content['technology_name']
+                technology_name=args['technology_name']
             )
-        elif content['content'] == 'subtechnology':
+        elif content_type == 'subtechnology':
+            table = all_content['Subtechnologies']
+            abort_if_exist(table, args['subtechnology_name'])
             new_content = SubTechnology(
-                technology_name=content['technology_name'],
-                subtechnology_name=content['subtechnology_name']
+                technology_name=args['technology_name'],
+                subtechnology_name=args['subtechnology_name']
             )
         else:
             return 'Wrong content key', 400
         db.session.add(new_content)
         db.session.commit()
-
         return marshal_wo_null(new_content), 200
 
 
@@ -96,7 +111,6 @@ class GetAllContent(Resource):
             'Technologies': technologies,
             'Subtechnologies': subtechnologies
         }
-
         return getall
 
 
@@ -106,16 +120,18 @@ class DeleteContent(Resource):
         content_type = args['content']
         content_id = args['id']
         got_content = globals().get(content_type).query.filter_by(id=content_id).first()
-        db.session.delete(got_content)
-        db.session.commit()
-
-        return f'Content from {content_type} with id {content_id} has been deleted'
+        if not got_content:
+            abort(404, message='Couldn\'t find requested content')
+        else:
+            db.session.delete(got_content)
+            db.session.commit()
+            return {'message': f'Content from {content_type} with id {content_id} has been deleted'}, 200
 
 
 # CHECK IF EVERYTHING IS WORKING THEN COMMIT
 api.add_resource(GetContent, '/get/')
 api.add_resource(PostContent, '/post/')
-api.add_resource(GetAllContent, '/getall/')
+api.add_resource(GetAllContent, '/get-all/')
 api.add_resource(DeleteContent, '/delete/')
 
 if __name__ == '__main__':
